move to devops OSX build (#2523)
* move to devops OSX build

* update cutil

* main() returns int

* no-deprecated-declarations

* test fix

+++ iothub_client/tests/iothub_client_properties_ut/main.c
@@ -7,5 +7,5 @@ int main(void)
 {
     size_t failedTestCount = 0;
     RUN_TEST_SUITE(iothub_client_properties_ut, failedTestCount);
-    return failedTestCount;
+    return (int)failedTestCount;
 }
