+++ provisioning_client/tests/hsm_client_tpm_ut/main.c
@@ -7,5 +7,5 @@ int main(void)
 {
     size_t failedTestCount = 0;
     RUN_TEST_SUITE(hsm_client_tpm_ut, failedTestCount);
-    return failedTestCount;
+    return (int)failedTestCount;
 }
